+++ src/temple_refactor3_tex/scripts/component/scroller.js
@@ -27,6 +27,21 @@ function initScroll() {
   ScrollTrigger.defaults({ scroller: pageConatainer }); //デフォルトのスクロール要素を設定
 
   const el = iNode.qs("[data-webgl]");
+
+  // const rect = el.getBoundingClientRect();
+  // const x = rect.left + 300;
+  // const pos = getWorldPosition({ left: x, width: rect.width }, canvasRect);
+
+  // gsap.to(os[0].mesh.position, {
+  //   x: pos.x,
+  //   scrollTrigger: {
+  //     trigger: el,
+  //     start: "center 70%",
+  //     end: "center center",
+  //     scrub: true,
+  //     pin: true,
+  //   },
+  // });
 }
 
 export default scroll;
