+++ src/temple_refactor3_tex/scripts/helper/index.js
