+++ src/temple_refactor3_tex/scripts/index.js
