+++ src/temple_refactor3_tex/scripts/vertex.glsl
