+++ src/temple_refactor3_tex/styles/test.scss
@@ -1,6 +1,5 @@
 body {
     margin: 0;
-    height: 200vh;
 }
 
 canvas {
