--- src/temple_refactor3rt/scripts/vertex1.glsl
@@ -1,6 +1,0 @@
-varying vec2 vUv;
-
-void main() {
-    vUv = uv;
-    gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-}
