+++ src/vertex.html
@@ -7,7 +7,7 @@
       Memorandum of understanding and practice journey when implementing
     </title>
     <!-- <link rel="stylesheet" href="style copy.css" /> -->
-    <script type="module" src="./temple_refactor3rt/scripts/index.js"></script>
+    <script type="module" src="./temple_refactor3_tex/scripts/index.js"></script>
   </head>
   <body>
     <div id="page-container">
